cattle_tracker/requirements.txt
Django==4.2.9
djangorestframework==3.14.0
django-filter==23.5
drf-spectacular==0.27.0
psycopg2-binary==2.9.9
gunicorn==21.2.0
python-decouple==3.8
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.1
Pillow==10.2.0
piexif==1.1.3
python-dotenv==1.0.0
whitenoise==6.6.0
ruff==0.1.15
pytest==7.4.4
pytest-django==4.8.0
factory-boy==3.3.0